fashion_segmentation/code/requirements.txt
matplotlib==3.1.1
PyYAML==5.1.2
Pillow>=8.1.1
opencv-python==4.1.1.26
scikit-learn==0.21.3
torch==1.2.0
torchvision==0.4.0
tqdm==4.36.1
flask
gunicorn
